bug: undefined userRole session variable
If the site was loaded for the very first time and the cache was cleared
the userRole session variable would be undefined and an error message
would appear on the page.  To fix this I check to see if userRole is
set.  If it's not i set it as Public.